+++ src/components/select/RadixSelect.tsx
@@ -245,7 +245,7 @@ const SelectContent = React.forwardRef<
         sideOffset={sideOffset}
         collisionPadding={8}
         className={cn(
-          'relative z-[100] min-w-[128px] overflow-y-auto overflow-x-hidden',
+          'relative z-[100] min-w-[128px] overflow-y-auto overflow-x-hidden scrollbar-thin',
           'bg-card border-default rounded-lg shadow-modal-sm',
           'data-[state=open]:animate-in data-[state=closed]:animate-out',
           'data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0',